+++ handlers/card.handler.go
@@ -2,39 +2,42 @@ package handlers
 
 import (
 	"br/simple-service/db/sqlc"
+	"errors"
 	"log"
 	"net/http"
 	"strconv"
 )
 
-func NewCardHandler(l *log.Logger, q *sqlc.Queries) *CardHandler {
+// TO DO : better db transaction
+
+func NewCardHandler(l *log.Logger, q *sqlc.Queries, u *AuthedUser) *CardHandler {
 	var c uint = 0
-	return &CardHandler{&Handler{l, q, &c}}
+	return &CardHandler{&Handler{l, q, &c, u}}
 }
 
 func (ch *CardHandler) CreateCardH(w http.ResponseWriter, r *http.Request) {
 	hp := HandlerParam{w, r, http.MethodPost, ch.createCard}
-	ch.h.handleRequest(hp)
+	ch.h.handleRequest(hp, ch.h.u)
 }
 
 func (ch *CardHandler) GetCardH(w http.ResponseWriter, r *http.Request) {
 	hp := HandlerParam{w, r, http.MethodGet, ch.getCard}
-	ch.h.handleRequest(hp)
+	ch.h.handleRequest(hp, ch.h.u)
 }
 
 func (ch *CardHandler) ListCardsH(w http.ResponseWriter, r *http.Request) {
 	hp := HandlerParam{w, r, http.MethodGet, ch.listCards}
-	ch.h.handleRequest(hp)
+	ch.h.handleRequest(hp, ch.h.u)
 }
 
 func (ch *CardHandler) UpdateCardH(w http.ResponseWriter, r *http.Request) {
 	hp := HandlerParam{w, r, http.MethodPut, ch.updateCard}
-	ch.h.handleRequest(hp)
+	ch.h.handleRequest(hp, ch.h.u)
 }
 
 func (ch *CardHandler) DeleteCardH(w http.ResponseWriter, r *http.Request) {
 	hp := HandlerParam{w, r, http.MethodDelete, ch.deleteCard}
-	ch.h.handleRequest(hp)
+	ch.h.handleRequest(hp, ch.h.u)
 }
 
 // the implementation
@@ -52,6 +55,18 @@ func (ch *CardHandler) createCard(w http.ResponseWriter, r *http.Request) error 
 		http.Error(w, "Invalid deck ID", http.StatusBadRequest)
 		return err
 	}
+
+	ownerID, err := ch.h.q.GetOwner(r.Context(), int32(deckID))
+	if err != nil {
+		http.Error(w, "No owner of the deck ID", http.StatusBadRequest)
+		return err
+	}
+
+	if ownerID != int32(ch.h.u.UserID) {
+		http.Error(w, "unauthorized", http.StatusNonAuthoritativeInfo)
+		return errors.New("unauthorized")
+	}
+
 	question := r.FormValue("question")
 	answer := r.FormValue("answer")
 
@@ -86,6 +101,17 @@ func (ch *CardHandler) getCard(w http.ResponseWriter, r *http.Request) error {
 		return err
 	}
 
+	ownerID, err := ch.h.q.GetOwner(r.Context(), int32(card.DeckID))
+	if err != nil {
+		http.Error(w, "No owner of the deck ID", http.StatusBadRequest)
+		return err
+	}
+
+	if ownerID != int32(ch.h.u.UserID) {
+		http.Error(w, "unauthorized", http.StatusNonAuthoritativeInfo)
+		return errors.New("unauthorized")
+	}
+
 	toJSON(w, card)
 	return nil
 }
@@ -95,6 +121,17 @@ func (ch *CardHandler) listCards(w http.ResponseWriter, r *http.Request) error {
 	if err != nil {
 		http.Error(w, "Invalid deck ID", http.StatusBadRequest)
 		return err
+	}
+
+	ownerID, err := ch.h.q.GetOwner(r.Context(), int32(deckID))
+	if err != nil {
+		http.Error(w, "No owner of the deck ID", http.StatusBadRequest)
+		return err
+	}
+
+	if ownerID != int32(ch.h.u.UserID) {
+		http.Error(w, "unauthorized", http.StatusNonAuthoritativeInfo)
+		return errors.New("unauthorized")
 	}
 
 	cards, err := ch.h.q.ListFlashcardsByDeck(r.Context(), int32(deckID))
@@ -120,6 +157,24 @@ func (ch *CardHandler) updateCard(w http.ResponseWriter, r *http.Request) error 
 		http.Error(w, "Invalid card ID", http.StatusBadRequest)
 		return err
 	}
+
+	deckID, err := ch.h.q.GetDeckID(r.Context(), int32(cardID))
+	if err != nil {
+		http.Error(w, "No deck associated with this card", http.StatusBadRequest)
+		return err
+	}
+
+	ownerID, err := ch.h.q.GetOwner(r.Context(), int32(deckID))
+	if err != nil {
+		http.Error(w, "No owner of the deck ID", http.StatusBadRequest)
+		return err
+	}
+
+	if ownerID != int32(ch.h.u.UserID) {
+		http.Error(w, "unauthorized", http.StatusNonAuthoritativeInfo)
+		return errors.New("unauthorized")
+	}
+
 	question := r.FormValue("question")
 	answer := r.FormValue("answer")
 
@@ -145,6 +200,23 @@ func (ch *CardHandler) deleteCard(w http.ResponseWriter, r *http.Request) error 
 	if err != nil {
 		http.Error(w, "Invalid card ID", http.StatusBadRequest)
 		return err
+	}
+
+	deckID, err := ch.h.q.GetDeckID(r.Context(), int32(cardID))
+	if err != nil {
+		http.Error(w, "No deck associated with this card", http.StatusBadRequest)
+		return err
+	}
+
+	ownerID, err := ch.h.q.GetOwner(r.Context(), int32(deckID))
+	if err != nil {
+		http.Error(w, "No owner of the deck ID", http.StatusBadRequest)
+		return err
+	}
+
+	if ownerID != int32(ch.h.u.UserID) {
+		http.Error(w, "unauthorized", http.StatusNonAuthoritativeInfo)
+		return errors.New("unauthorized")
 	}
 
 	card, err := ch.h.q.DeleteFlashcard(r.Context(), int32(cardID))
